Create a decorator for Notification
As this project is a gem only, it does not have a complete Notification model. We have a PaymentEngines that brings us the Payment instance with notification that should be used for implement our processing.

The decorator is needed to avoid missing implementations for Payment's notification instance, then avoiding to crash the application.